# openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:4000/apaman198/"
    },
    {
      "url": "https://localhost:4000/apaman198/"
    }
  ],
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0-oas3",
    "title": "Montaman API Document",
    "termsOfService": "https://monta.ir/apaman",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/acc/payType": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Get the list of Payment Type",
        "description": "show Payment type",
        "operationId": "paymentTypeShow",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnumType"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Failed!!!"
          },
          "500": {
            "description": "server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/acc/bank": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "list of bank name",
        "description": "list of Bank name",
        "operationId": "enumBank",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnumType"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Failed!!!"
          },
          "500": {
            "description": "server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "EnumType": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "PersianName": {
            "type": "string"
          }
        }
      },
...
    }
  }
}